Created Loggify decorators
