+++ src/models/block.ts
@@ -7,6 +7,7 @@ import { BitcoinBlockType, BlockHeaderObj } from '../types/Block';
 import { ChainNetwork } from '../types/ChainNetwork';
 import { TransformableModel } from '../types/TransformableModel';
 import logger from '../logger';
+import { LoggifyMethod, LoggifyFunction } from '../decorators/Loggify';
 const async = require('async');
 
 export interface IBlock {
@@ -73,7 +74,7 @@ BlockSchema.index({ chain: 1, network: 1, processed: 1, height: -1 });
 BlockSchema.index({ chain: 1, network: 1, timeNormalized: 1 });
 BlockSchema.index({ previousBlockHash: 1 });
 
-BlockSchema.statics.addBlock = function(
+BlockSchema.statics.addBlock = LoggifyFunction(function(
   params: AddBlockParams,
   callback: CallbackType
 ) {
@@ -165,7 +166,7 @@ BlockSchema.statics.addBlock = function(
       );
     }
   );
-};
+}, 'BlockSchema.addBlock');
 
 BlockSchema.statics.getPoolInfo = function(coinbase: string) {
   //TODO need to make this actually parse the coinbase input and map to miner strings
